2017-11-8

+++ php/ins_php.sh
@@ -12,8 +12,8 @@ yum -y install epel-release
 yum -y install gcc-c++ make libxml2-devel curl-devel libjpeg-devel libpng-devel freetype-devel zlib-devel pcre-devel openssl-devel libxml2 libxml2-devel openssl openssl-devel libcurl libcurl-devel libjpeg libjpeg-devel freetype freetype-devel libc-client libc-client-devel libtool.x86_64 libtool-ltdl.x86_64 libtool-ltdl-devel.x86_64 openldap-devel libjpeg.x86_64 libtool libpng.x86_64 freetype.x86_64 libjpeg-devel.x86_64 libpng-devel.x86_64 freetype-devel.x86_64 krb5* mcrypt mhash-devel libmcrypt-devel mysql-devel
 /usr/bin/id www || useradd -M www
 
-ln -sf /usr/lib64 /usr/kerberos/lib
-ln -sf /usr/lib64/libc-client.so /usr/lib/libc-client.so
+[ -d "/usr/lib64 /usr/kerberos/lib" ] && ln -sf /usr/lib64 /usr/kerberos/lib
+[ -f "/usr/lib64/libc-client.so /usr/lib/libc-client.so" ] && ln -sf /usr/lib64/libc-client.so /usr/lib/libc-client.so
